requirements.txt
numpy==1.19.5
opencv-python==4.5.5.62
Pillow==8.4.0
torch==1.10.1
torchvision==0.11.2
einops==0.4.0
